+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 -   Bumped TMC-langs version to 0.36.2
 -   .csproj files are no longer included when submitting C# exercises
+-   Set the minimum supported VSCode version to 1.86.0 (released in January 2024)
+-   Fixed the TMC-langs CLI's permissions being set to 111
 
 ## [3.0.3] - 2024-02-26